refactor: use function to make test cases reusable
